# [LinearSystem] Optim: Only account for affected DoFs