--- a/Sofa/Component/LinearSystem/src/sofa/component/linearsystem/MatrixLinearSystem.inl
+++ b/Sofa/Component/LinearSystem/src/sofa/component/linearsystem/MatrixLinearSystem.inl
@@ -757,12 +757,12 @@ void MatrixLinearSystem<TMatrix, TVector>::projectMappedMatrices(const core::Mec
             continue;
         }
 
-        const MappingJacobians<JacobianMatrixType> J0 = computeJacobiansFrom(pair[0], mparams);
-        const MappingJacobians<JacobianMatrixType> J1 = computeJacobiansFrom(pair[1], mparams);
+        LocalMappedMatrixType<Real>* crs = mappedMatrix.get();
+
+        const MappingJacobians<JacobianMatrixType> J0 = computeJacobiansFrom(pair[0], mparams, crs);
+        const MappingJacobians<JacobianMatrixType> J1 = computeJacobiansFrom(pair[1], mparams, crs);
 
         const sofa::type::fixed_array<MappingJacobians<JacobianMatrixType>, 2> mappingMatricesMap { J0, J1 };
-
-        LocalMappedMatrixType<Real>* crs = mappedMatrix.get();
 
         sofa::component::linearsystem::addMappedMatrixToGlobalMatrixEigen(
             pair, crs, mappingMatricesMap, m_mappingGraph, this->getSystemMatrix());
@@ -770,7 +770,31 @@ void MatrixLinearSystem<TMatrix, TVector>::projectMappedMatrices(const core::Mec
 }
 
 template <class TMatrix, class TVector>
-auto MatrixLinearSystem<TMatrix, TVector>::computeJacobiansFrom(BaseMechanicalState* mstate, const core::MechanicalParams* mparams)
+std::vector<unsigned int> MatrixLinearSystem<TMatrix, TVector>::identifyAffectedDoFs(BaseMechanicalState* mstate, LocalMappedMatrixType<Real>* crs)
+{
+    const auto blockSize = mstate->getMatrixBlockSize();
+    std::set<unsigned int> setAffectedDoFs;
+
+    for (std::size_t it_rows_k = 0; it_rows_k < crs->rowIndex.size(); it_rows_k++)
+    {
+        typename LocalMappedMatrixType<Real>::Range rowRange(crs->rowBegin[it_rows_k], crs->rowBegin[it_rows_k + 1]);
+        for (auto xj = rowRange.begin(); xj < rowRange.end(); ++xj) // for each non-null block
+        {
+            const sofa::SignedIndex col = crs->colsIndex[xj];
+            const auto k = crs->colsValue[xj];
+            if (k != static_cast<Real>(0))
+            {
+                const sofa::SignedIndex dofId = col / blockSize;
+                setAffectedDoFs.insert(dofId);
+            }
+        }
+    }
+
+    return std::vector( setAffectedDoFs.begin(), setAffectedDoFs.end() );
+}
+
+template <class TMatrix, class TVector>
+auto MatrixLinearSystem<TMatrix, TVector>::computeJacobiansFrom(BaseMechanicalState* mstate, const core::MechanicalParams* mparams, LocalMappedMatrixType<Real>* crs)
 -> MappingJacobians<JacobianMatrixType>
 {
     auto cparams = core::ConstraintParams(*mparams);
@@ -786,9 +810,10 @@ auto MatrixLinearSystem<TMatrix, TVector>::computeJacobiansFrom(BaseMechanicalSt
 
     auto mappingJacobianId = sofa::core::MatrixDerivId::mappingJacobian();
 
-    sofa::type::vector<unsigned int> listAffectedDoFs(mstate->getSize());
-    std::iota(listAffectedDoFs.begin(), listAffectedDoFs.end(), 0);
-    mstate->buildIdentityBlocksInJacobian(listAffectedDoFs, mappingJacobianId);
+    {
+        const std::vector<unsigned> listAffectedDoFs = identifyAffectedDoFs(mstate, crs);
+        mstate->buildIdentityBlocksInJacobian(listAffectedDoFs, mappingJacobianId);
+    }
 
     const auto parentMappings = getMappingGraph().getBottomUpMappingsFrom(mstate);
     for (auto* mapping : parentMappings)