@alxbilger 2023-07-28
I confirm that, in the `caduceus.scn`, the jacobian matrices of the `BarycentricMapping`s were built even if the mapping does not contribute to the stiffness matrix. That is why, this PR speeds up the caduceus: no more unnecessary construction of the jacobian matrices.