@alxbilger 2023-07-28
Latest benchmark on the caduceus:

Before
```
[BatchGUI] 10000 iterations done in 37.8088 s ( 264.489 FPS).
```

After
```
[BatchGUI] 10000 iterations done in 14.4789 s ( 690.661 FPS).
```